requirements.txt
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
